fix: pull data from url instead of file system
