+++ js/examples/write_html.js
